Abstract GitHub interactions
This will add a layer of abstraction around the GitHub API interactions
that write to the API (i.e., stuff affected by `DRY_RUN`.  This helps
test the tool without actually writing to GitHub.

An additional logging variable was added, `LOG_LEVEL` that allows the
user to specify how verbose (higher = less verbose) the tool will be
when determing which messages to display.  The default is 20 (i.e.,
INFO); 10 would correspond to DEBUG while 50 would correspond to
CRITICAL.

The documentation and sample env file were updated.